Move `FlexSplitColumn` and `BlankHeader` to `devtools_app_shared`
