+++ packages/devtools_app/lib/src/shared/common_widgets.dart
@@ -673,22 +673,6 @@ abstract class ScaffoldAction extends StatelessWidget {
   }
 }
 
-/// A blank, drop-in replacement for [AreaPaneHeader].
-///
-/// Acts as an empty header widget with zero size that is compatible with
-/// interfaces that expect a [PreferredSizeWidget].
-class BlankHeader extends StatelessWidget implements PreferredSizeWidget {
-  const BlankHeader({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-
-  @override
-  Size get preferredSize => Size.zero;
-}
-
 /// Button to open related information / documentation.
 ///
 /// [tooltip] specifies the hover text for the button.
@@ -1230,18 +1214,6 @@ class _BreadcrumbPainter extends CustomPainter {
   }
 }
 
-class JsonViewer extends StatefulWidget {
-  const JsonViewer({
-    super.key,
-    required this.encodedJson,
-  });
-
-  final String encodedJson;
-
-  @override
-  State<JsonViewer> createState() => _JsonViewerState();
-}
-
 /// A wrapper for a Text widget, which allows for concatenating text if it
 /// becomes too long.
 class TextViewer extends StatelessWidget {
@@ -1272,6 +1244,18 @@ class TextViewer extends StatelessWidget {
       style: style,
     );
   }
+}
+
+class JsonViewer extends StatefulWidget {
+  const JsonViewer({
+    super.key,
+    required this.encodedJson,
+  });
+
+  final String encodedJson;
+
+  @override
+  State<JsonViewer> createState() => _JsonViewerState();
 }
 
 class _JsonViewerState extends State<JsonViewer>
