+++ packages/devtools_app/test/inspector/layout_explorer/inspector_data_models_test.dart
@@ -112,7 +112,7 @@ void main() {
     group('childrenRenderProperties tests', () {
       const maxMainAxisDimension = 500.0;
 
-      double maxSizeAvailable(Axis axis) => maxMainAxisDimension;
+      double maxSizeAvailable(Axis _) => maxMainAxisDimension;
 
       List<RenderProperties> childrenRenderProperties(
         FlexLayoutProperties properties,
