+++ packages/devtools_app_shared/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.10-wip
 * Add `DTDManager` class and export from `service.dart`.
 * Add `showDevToolsDialog` helper method.
+* Add `FlexSplitColumn` and `BlankHeader` common widgets.
 
 ## 0.0.9
 * Bump `package:web` to `^0.4.1`.
